Updates Repository name
