+++ account/src/main/java/com/user/account/services/TransactionService.java
@@ -1,0 +1,12 @@
+package com.user.account.services;
+
+import java.util.ArrayList;
+
+import com.user.account.entity.Transactions;
+
+public interface TransactionServices {
+	
+	public Transactions addTransaction(Transactions newTransaction);
+	
+	public ArrayList< Transactions> getSummary(long accountNumber);
+}
