--- account/src/main/java/com/user/account/services/TransactionServices.java
@@ -1,12 +1,0 @@
-package com.user.account.services;
-
-import java.util.ArrayList;
-
-import com.user.account.entity.Transactions;
-
-public interface TransactionServices {
-	
-	public Transactions addTransaction(Transactions newTransaction);
-	
-	public ArrayList< Transactions> getSummary(long accountNumber);
-}
